requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
beautifulsoup4==4.13.4
    # via bs4
bs4==0.0.2
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
fastapi==0.115.14
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
psycopg2-binary==2.9.10
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-multipart==0.0.20
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
requests==2.32.4
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
starlette==0.46.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via tt-fastapi-file-metadata-parser (pyproject.toml)
